+++ .gitignore
@@ -1,2 +1,3 @@
 /.idea/
 /out/
+/Electricidad-Chispas.iml
