+++ src/com/iesam/chispas/domain/models/Client.java
@@ -1,3 +1,5 @@
+package com.iesam.chispas.domain.models;
+
 public interface Client {
 
     public String getName();
