+++ src/com/iesam/chispas/domain/models/Product.java
@@ -1,18 +1,40 @@
-public class Service implements Sale{
+package com.iesam.chispas.domain.models;
 
-    private String productService;
+public class Product implements Sale {
+
+    private String productCode;
     private String name;
+    private String brand;
+    private String model;
     private Integer price;
     private String iva;
 
+    //metodos
+
+    public String getBrand() {
+        return brand;
+    }
+
+    public void setBrand(String brand) {
+        this.brand = brand;
+    }
+
+    public String getModel() {
+        return model;
+    }
+
+    public void setModel(String model) {
+        this.model = model;
+    }
+
     @Override
     public String getCode() {
-        return productService;
+        return productCode;
     }
 
     @Override
     public void setCode(String code) {
-        this.productService = code;
+        this.productCode = code;
     }
 
     @Override
@@ -22,7 +44,7 @@ public class Service implements Sale{
 
     @Override
     public void setName(String name) {
-        this.name= name;
+        this.name = name;
     }
 
     @Override
@@ -42,6 +64,6 @@ public class Service implements Sale{
 
     @Override
     public void setIva(String iva) {
-        this.iva= iva;
+            this.iva = iva;
     }
 }
