+++ src/com/iesam/chispas/domain/models/Service.java
@@ -1,38 +1,20 @@
-public class Product implements Sale{
+package com.iesam.chispas.domain.models;
 
-    private String productCode;
+public class Service implements Sale {
+
+    private String productService;
     private String name;
-    private String brand;
-    private String model;
     private Integer price;
     private String iva;
 
-    //metodos
-
-    public String getBrand() {
-        return brand;
-    }
-
-    public void setBrand(String brand) {
-        this.brand = brand;
-    }
-
-    public String getModel() {
-        return model;
-    }
-
-    public void setModel(String model) {
-        this.model = model;
-    }
-
     @Override
     public String getCode() {
-        return productCode;
+        return productService;
     }
 
     @Override
     public void setCode(String code) {
-        this.productCode = code;
+        this.productService = code;
     }
 
     @Override
@@ -42,7 +24,7 @@ public class Product implements Sale{
 
     @Override
     public void setName(String name) {
-        this.name = name;
+        this.name= name;
     }
 
     @Override
@@ -62,6 +44,6 @@ public class Product implements Sale{
 
     @Override
     public void setIva(String iva) {
-            this.iva = iva;
+        this.iva= iva;
     }
 }
